Improved teleport
added mask and layer collide which made it possible to teleport trough enemy but not walls. This is tough only teleport x axis but should not be a problem adding y axis also. changed the floor so its not tilted. Maybe fix so insteed of error when its a "block" its teleport to the closets place like you teleport right before the wall.